Move the solver to game class and add clear screen actions.
